Issue #30: Updated citation extraction script
- updated gitignore to ignore txt and csv files from scripts directory
- split the citation extraction script into smaller scripts
- added a bash script to run the python script in backgorund
