+++ scripts/clean.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+
+# This Source Code Form is subject to the terms of the MIT
+# License. If a copy of the same was not distributed with this
+# file, You can obtain one at
+# https://github.com/akhilpandey95/scholarlyimpact/blob/master/LICENSE.
+
+import os
+
+os.remove('altmetric_ids.txt')
+os.remove('citations_dim.csv')
